Refactor tests for constraints-violations a bit (new package)
